+++ src/main/java/org/prgms/voucherProgram/domain/wallet/controller/WalletController.java
@@ -12,7 +12,6 @@ import org.prgms.voucherProgram.domain.voucher.service.VoucherService;
 import org.prgms.voucherProgram.domain.wallet.dto.WalletRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -91,11 +90,5 @@ public class WalletController {
         Customer customer = voucherService.findCustomer(voucherId);
         model.addAttribute("customer", customer);
         return "wallet/customer-show";
-    }
-
-    @ExceptionHandler
-    public String handleException(Exception e, Model model) {
-        model.addAttribute("error", e.getMessage());
-        return "error";
     }
 }
