+++ src/main/java/org/prgms/voucherProgram/global/exception/ExceptionControllerAdvice.java
@@ -1,31 +1,26 @@
 package org.prgms.voucherProgram.global.exception;
 
-import java.util.Date;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@RestControllerAdvice
+@ControllerAdvice
 public class ExceptionControllerAdvice {
 
     private static final Logger logger = LoggerFactory.getLogger(ExceptionControllerAdvice.class);
 
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<Object> handleCustomException(Exception e) {
-        ExceptionResponse exceptionResponse = new ExceptionResponse(HttpStatus.BAD_REQUEST.value(), new Date(),
-            e.getMessage());
-        return new ResponseEntity<>(exceptionResponse, HttpStatus.BAD_REQUEST);
+    @ExceptionHandler
+    public String handleCustomerException(IllegalArgumentException e, Model model) {
+        model.addAttribute("error", e.getMessage());
+        return "error";
     }
 
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Object> handleAllException(Exception e) {
+    @ExceptionHandler
+    public String handleAllException(Exception e, Model model) {
         logger.error(e.getMessage());
-        ExceptionResponse exceptionResponse = new ExceptionResponse(HttpStatus.BAD_REQUEST.value(), new Date(),
-            e.getMessage());
-        return new ResponseEntity<>(exceptionResponse, HttpStatus.BAD_REQUEST);
+        model.addAttribute("error", e.getMessage());
+        return "error";
     }
 }
